pulling only fields out of common-functions to use in cmp lib as consumer-self-test
